move deviceType to the same file
