--- src/types/deviceState.ts
@@ -1,7 +1,0 @@
-export type DeviceState = {
-  isSupported: boolean;
-  deviceType: "Mobile" | "Tablet" | "Desktop";
-  isMobile: boolean;
-  isTablet: boolean;
-  isDesktop: boolean;
-};
